+++ Documentation/Changelog.md
@@ -7,6 +7,7 @@ This version is currently in beta.
 - Added support for .NET Core 3.0 target framework
 - Introduced `Sequential` and `DebuggableParallel` primitives in `Gapotchenko.FX.Threading` module.
 Both primitives form a drop-in replacement for `System.Threading.Tasks.Parallel` and are useful for debugging purposes
+- Added `CommandLine.OemEncoding` property that gets OEM encoding used by Windows command line and console applications
 - Implemented polyfill for `Enumerable.ToHashSet<T>` operation
 - Implemented polyfills for `BitConverter.SingleToInt32Bits` and `Int32BitsToSingle` operations
 - `Empty.Task` is not suggested by the code editor when it is natively provided by the host platform